some performance work. user more Trove, store frame data in ThreadLocal instead of dynamic var

+++ src/java/noria/LCS.java
@@ -4,7 +4,7 @@ import java.util.*;
 
 public class LCS {
 
-    public static int[] lcs(int[] a, int b[]) {
+    public static long[] lcs(long[] a, long b[]) {
         int[][] lengths = new int[a.length+1][b.length+1];
  
         // row 0 and column 0 are initialized to 0 already
@@ -17,7 +17,7 @@ public class LCS {
                     lengths[i+1][j+1] = len = Math.max(lengths[i+1][j], lengths[i][j+1]);
  
         // read the substring out from the matrix
-        int[] result = new int[len];
+        long[] result = new long[len];
         for (int x = a.length, y = b.length, i = 0;
              x != 0 && y != 0;) {
             if (lengths[x][y] == lengths[x-1][y])
@@ -33,4 +33,3 @@ public class LCS {
         return result;
     }
 }
-
